Seperate admin page to components
